+++ app/src/main/res/layout/make_meme.xml
@@ -6,110 +6,4 @@
     android:orientation="vertical">
 
 
-    <LinearLayout
-        android:id="@+id/tlay"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:orientation="vertical">
-
-        <EditText
-            android:id="@+id/edit_text1"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:background="#000000"
-            android:hint="Title"
-
-            android:textColor="#ffffff"
-            android:textColorHint="#ffffff" />
-
-        <EditText
-            android:id="@+id/edit_text2"
-
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:background="#000000"
-            android:hint="Subtitle"
-            android:textColor="#ffffff"
-            android:textColorHint="#ffffff" />
-
-        <Button
-            android:id="@+id/go"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:text="MEME" />
-    </LinearLayout>
-
-
-    <LinearLayout
-        android:id="@+id/blay"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_alignParentBottom="true"
-        android:layout_centerHorizontal="true"
-        android:orientation="horizontal"
-        android:weightSum="3">
-
-        <Button
-            android:id="@+id/save"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_weight="1"
-
-            android:text="SAVE" />
-
-        <Button
-            android:id="@+id/load"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_weight="1"
-            android:text="LOAD" />
-
-        <Button
-            android:id="@+id/share"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_weight="1"
-            android:text="SHARE" />
-    </LinearLayout>
-
-    <RelativeLayout
-        android:id="@+id/lay"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:layout_above="@+id/blay"
-        android:layout_below="@+id/tlay"
-        android:layout_centerInParent="true">
-
-        <ImageView
-            android:id="@+id/imageView"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:layout_above="@+id/textView1"
-            android:padding="20dp"
-            android:scaleType="fitXY" />
-
-
-        <TextView
-            android:id="@+id/textView1"
-
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_above="@+id/textView2"
-            android:layout_centerHorizontal="true"
-            android:gravity="center"
-            android:textColor="#ffffff"
-            android:textSize="40dp" />
-
-        <TextView
-            android:id="@+id/textView2"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_alignParentBottom="true"
-            android:layout_centerHorizontal="true"
-            android:gravity="center"
-            android:textColor="#ffffff"
-            android:textSize="20dp" />
-
-    </RelativeLayout>
-
 </RelativeLayout>
